[FE] REFACTOR: toggle Menu 조건 최적화 및 파일 하나로 묶음 #817
